修复 vs.net 编译警告

+++ fibjs/include/parse.h
@@ -20,7 +20,7 @@ public:
         string(s), pos(0)
     {
         if (n < 0)
-            n = qstrlen(s);
+            n = (int)qstrlen(s);
         sz = n;
     }
